# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAtext API
  description: REST API specification for WASAtext messaging system.
  version: "3.0.1"
servers:
  - url: http://localhost:3000
    description: Test server running locally, for development purposes.
...
paths:
...
  /messages:
    post:
      tags: [messages]
      summary: Send message
      description: Sends a new message to either a group conversation or private conversation. If conversation doesn't exist between users and no conversationId is provided, creates a new private conversation. Optionally supports replying to existing messages.
      operationId: sendMessage
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - senderUsername
                - content
              properties:
                senderUsername:
                  type: string
                  description: Username of the sender
                  minLength: 1
                  maxLength: 64
                  pattern: "^[a-zA-Z0-9_-]+$"
                  example: "Karol"
                receiverUsername:
                  type: string
                  description: Username of the receiver (required for private conversations, optional for groups)
                  minLength: 1
                  maxLength: 64
                  pattern: "^[a-zA-Z0-9_-]+$"
                  example: "alice"
                conversationId:
                  type: string
                  description: Conversation ID (optional, used for group messages or existing conversations)
                  minLength: 1
                  maxLength: 64
                  pattern: "^[a-zA-Z0-9_-]+$"
                  example: "12345"
                content:
                  type: string
                  description: Message content
                  minLength: 1
                  maxLength: 500
                  example: "Hello, whats up?"
                replyingToId:
                  type: string
                  description: ID of the message being replied to (optional, for threading)
                  minLength: 1
                  maxLength: 64
                  pattern: "^[0-9]+$"
                  example: "98765"
                attachments:
                  type: array
                  description: Optional file attachments
                  minItems: 0
                  maxItems: 10
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        "400":
          description: Invalid request (missing required fields or conversation not found)
        "401":
          description: Unauthorized - invalid or missing token
        "500":
          description: Server error
  /messages/{messageId}:
    delete:
      tags: [messages]
      summary: Delete message
      description: Deletes a message. Only the sender can delete their own message.
      operationId: deleteMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        "204":
          description: Message deleted successfully
        "401":
          description: Unauthorized - invalid token or not the message sender
        "404":
          description: Message not found
        "500":
          description: Server error
  /messages/{messageId}/status:
    patch:
      tags: [messages]
      summary: Update message status
      description: Updates the status of a message (e.g., from 'sent' to 'delivered' or 'received'). User must be a member of the conversation but not the sender of the message.
      operationId: updateMessageStatus
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageStatusUpdate'
      responses:
        "204":
          description: Message status updated successfully
        "400":
          description: Invalid request body or invalid status value
        "401":
          description: Unauthorized - invalid or missing token
        "403":
          description: Forbidden - user not in conversation or is the message sender
        "404":
          description: Message not found
        "500":
          description: Server error
  /messages/{messageId}/forwards:
    post:
      tags: [messages]
      summary: Forward message
      description: Forwards an existing message to another conversation. User must be a member of the target conversation.
      operationId: forwardMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardRequest'
      responses:
        "201":
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty response on success
        "400":
          description: Invalid request body
        "401":
          description: Unauthorized - invalid or missing token
        "403":
          description: Forbidden - user not in target conversation
        "404":
          description: Original message not found
        "500":
          description: Server error
  /attachments/{messageId}:
    post:
      tags: [messages]
      summary: Get attachment from message
      description: Retrieves a specific attachment from a message.
      operationId: getAttachmentFromMessage
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        "200":
          description: Attachment retrieved successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
        "404":
          description: Message or attachment not found
        "500":
          description: Server error
  /messages/{messageId}/attachments:
    get:
      tags: [messages]
      summary: Get message attachments
      description: Retrieves all attachments for a message.
      operationId: getMessageAttachments
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        "200":
          description: Attachments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 100
                items:
                  type: object
                  properties:
                    content:
                      type: string
                      description: Base64 encoded attachment content
                      minLength: 0
                      maxLength: 10485760
        "400":
          description: Invalid message ID
        "401":
          description: Unauthorized
        "404":
          description: Message not found
        "500":
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MessageId:
      name: messageId
      in: path
      required: true
      description: Unique message identifier
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_-]+$"
...
  schemas:
...
    ForwardRequest:
      type: object
      description: Forward message request. Must provide either addressingConversationID (for existing conversation) or newContactUsername (to create new private conversation).
      required: [username]
      properties:
        addressingConversationID:
          type: integer
          description: Target conversation ID to forward the message to (optional, use when forwarding to existing conversation)
          minimum: 1
          example: 12345
        newContactUsername:
          type: string
          description: Username to forward message to as a new contact (optional, use when forwarding to new conversation)
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "alice"
        username:
          type: string
          description: Username of the user forwarding the message
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "Karol"
...
    MessageResponse:
      type: object
      description: Response after sending or reacting to a message
      required: [id, senderId, content, conversationId, timestamp]
      properties:
        id:
          type: string
          description: Message ID
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "54321"
        senderId:
          type: string
          description: Username of the sender
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "Karol"
        content:
          type: string
          description: Message content (or emoji for reactions)
          minLength: 1
          maxLength: 500
          example: "Hello, whats up?"
        conversationId:
          type: string
          description: Conversation ID
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "12345"
        timestamp:
          type: string
          format: date-time
          description: Message timestamp
          example: "2026-01-01T21:37:00Z"
...
    MessageStatusUpdate:
      type: object
      description: Message status update payload
      required: [username, status]
      properties:
        username:
          type: string
          description: Username of the user updating the status
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "Karol"
        status:
          type: string
          description: New message status (must be one of 'sent', 'delivered', 'received', 'hidden')
          minLength: 4
          maxLength: 9
          pattern: "^(sent|delivered|received|hidden)$"
          enum: [sent, delivered, received, hidden]
          example: "received"